Allow user to configure link viewer
Following the pattern of image-viewer, video-viewer, and audio-viewer, this adds an entry to the [media] section of the config file called link-viewer, which allows the user to set his own browser.  This is important because, as a TUI application, tut may be running with no display, in which case the usual browser invoked by xdg-open is likely unavailable but text-mode browsers may be.